requirements.txt
pytest
pytest-bdd == 3.4.0
allure-pytest
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.8.0-cp39-cp39-win_amd64.whl